refactor at github_user_scraper to support update method

+++ app/helpers/users_helper.rb
@@ -68,6 +68,6 @@ module UsersHelper
   def user_page_btns
     link_to("Update GH info", update_gh_info_path(@user), class: "btn btn-outline-success") +
     link_to("Edit user", edit_user_path(@user), class: "btn btn-outline-secondary") +
-    link_to("Delete user", user_path(@user), method: :delete, class: "btn btn-outline-danger")
+    link_to("Delete user", user_path(@user), method: :delete, class: "btn btn-outline-danger", data: { confirm: 'Are you sure?' })
   end
 end
